+++ CHANGELOG.md
@@ -12,12 +12,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `exh.users.globalRoles` now has `find`, `findFirst`, `findById` and `findByName` methods
 - `exh.users.groupRoles` now has `find`, `findFirst`, `findById` and `findByName` methods
 - `exh.auth.applications` now has `find`, `findFirst`, `findById` and `findByName` methods
+- `exh.auth.oauth1.tokens` now has `find`, `findFirst`, `findById` and `remove` methods
+- `exh.auth.oauth2.tokens` now has `find`, `findFirst`, `findById` and `remove` methods
 - The `exh.events.subscriptions` methods now also support the new `retriable` field
 
 ### Deprecated
 - `exh.users.globalRoles.get` is deprecated in favor of `exh.users.globalRoles.find`
 - `exh.users.groupRoles.get` is deprecated in favor of `exh.users.groupRoles.find`
 - `exh.auth.applications.get` is deprecated in favor of `exh.auth.applications.find`
+- `exh.auth.oauth1.getTokens` is deprecated in favor of `exh.auth.oauth1.tokens.find`
+- `exh.auth.oauth1.removeToken` is deprecated in favor of `exh.auth.oauth1.tokens.remove`
 - `exh.events.subscriptions.create` should no longer be used, this is managed by the system
 
 ### Fixed